Re-add documentation, remove exploratory files
