+++ src/Symbolize.hs
@@ -3,19 +3,16 @@
 -- Symbols, also known as Atoms or Interned Strings, are a common technique
 -- to reduce memory usage and improve performance when using many small strings.
 --
--- By storing a single copy of each encountered string in a global table and giving out indexes to that table,
+-- By storing a single copy of each encountered string in a global table and giving out pointers to the stored keys,
 -- it is possible to compare strings for equality in constant time, instead of linear (in string size) time.
 --
--- The main advantages of Symbolize over existing symbol table implementations are:
+-- The main advantages of Symbolize over other symbol table implementations are:
 --
 -- - Garbage collection: Symbols which are no longer used are automatically cleaned up.
--- - `Symbol`s have a small memory footprint (a pinned `ByteArray#`) unpacked by GHC.
--- - Support for any `Textual` type, including `String`, (strict and lazy) `Data.Text`, (strict and lazy) `Data.ByteString` etc.
+-- - `Symbol`s have a memory footprint of exactly 1 Word and are nicely unpacked by GHC.
+-- - Support for any `Textual` type, including `String`, (strict and lazy) `Data.Text`, (strict and lazy) `Data.ByteString`, `ShortText`, `ShortByteString`, etc.
 -- - Thread-safe.
--- - Efficient: 
---    - `unintern` only has to follow a pointer
---    - `lookup` reads an IOref but is free of atomic memory barries
--- - Support for as much symbols as you have memory.
+-- - Efficient: Calls to `lookup` and `unintern` are free of atomic memory barriers (and never have to wait on a concurrent thread running `intern`)
 --
 -- == Basic usage
 --
@@ -64,9 +61,12 @@
 -- Just (Symbolize.intern "hello")
 --
 -- Symbols make great keys for `Data.HashMap` and `Data.HashSet`.
--- Hashing them is a no-op and they are guaranteed to be unique.
--- _(Note that hash order changes between program runs, since
--- the pointer address of the pinned text bytes is internally used.)_
+-- Hashing them is a no-op and they are guaranteed to be unique:
+--
+-- >>> Data.Hashable.hash hello
+-- 1
+-- >>> fmap Data.Hashable.hash niceCheeses
+-- [2,3,4]
 --
 -- For introspection, you can look at how many symbols currently exist:
 --
@@ -91,91 +91,119 @@
 {-# LANGUAGE GHC2021 #-}
 {-# LANGUAGE MagicHash #-}
 {-# LANGUAGE DerivingStrategies #-}
-module Symbolize (
-    -- * Symbol
-    Symbol, 
-    intern, 
-    unintern, 
-    lookup, 
-    Textual (..), 
+{-# LANGUAGE UnliftedNewtypes #-}
+{-# LANGUAGE UnboxedTuples #-}
+{-# LANGUAGE UnliftedDatatypes #-}
+module Symbolize   
+( -- * Symbol
+    Symbol,
+    intern,
+    unintern,
+    lookup,
+    Textual (..),
 
     -- * Introspection & Metrics
-    GlobalSymbolTable, 
-    globalSymbolTable, 
-    globalSymbolTableSize
-    )
-    where
-
+    GlobalSymbolTable,
+    globalSymbolTable,
+    globalSymbolTableSize,
+  ) where
 import Prelude hiding (lookup)
-import Control.Applicative ((<|>))
-import Data.Function ((&))
-import Data.String (IsString (..))
-import Data.Text.Short (ShortText)
-import Data.Text.Short qualified as Text.Short
-import Data.Text.Short.Unsafe qualified as Text.Short.Unsafe
---import Data.Text.Display (Display (..))
-import Data.ByteString.Short (ShortByteString(..))
-import Data.Primitive.ByteArray (ByteArray(..), ByteArray#)
-import Data.Hashable qualified as Hashable
-import Control.DeepSeq (NFData (..))
-import GHC.Int (Int(I#))
-import GHC.Exts (addr2Int#, byteArrayContents#)
-import GHC.Read (Read (..))
-import Text.Read (Lexeme (Ident), lexP, parens, prec, readListPrecDefault)
-import qualified Text.Read
+import GHC.Exts (reallyUnsafePtrEquality#, UnliftedType,  Weak#, mkWeak#, makeStableName#, stableNameToInt#, Int(I#), deRefWeak#, ByteArray#)
+import GHC.IO (IO(IO))
 import qualified System.IO.Unsafe
 import Data.IORef (IORef)
 import qualified Data.IORef as IORef
-import System.Mem.Weak (Weak)
-import qualified System.Mem.Weak as Weak
+import Data.ByteString.Short (ShortByteString(SBS))
+import Data.Text.Short (ShortText)
+import qualified Data.Text.Short as Text.Short
+import qualified Data.Text.Short.Unsafe as Text.Short.Unsafe
 import Data.Map.Strict (Map)
-import qualified Data.Map.Strict as Map
-import Data.Set (Set)
-import qualified Data.Set as Set
+import Data.Map.Strict qualified as Map
+import Control.DeepSeq (NFData(..))
+import Data.Hashable (Hashable(..))
+import Data.String (IsString(..))
+import qualified Text.Read
+import GHC.Read (Read (..))
+import Text.Read (Lexeme (Ident), lexP, parens, prec, readListPrecDefault)
+import Control.Applicative ((<|>))
 
-
-import Symbolize.Accursed
 import Symbolize.Textual (Textual)
 import Symbolize.Textual qualified as Textual
+import Symbolize.Accursed qualified
 
--- INVARIANTS:
+-- | A string-like type with O(1) equality and comparison.
 --
--- - The ByteArray is a valid UTF-8 string and thus can be cheaply turned back into a 
---   `ShortText` simply by casting.
--- - The ByteArray is pinned, because we rely on its pointer address being stable.
-newtype Symbol = Symbol ByteArray
+-- A Symbol represents a string (any `Textual`, so String, Text, ShortText, ByteString, ShortByteString, etc.)
+-- However, the Symbol itself only is one word large: the pointer to an underlying `ByteArray#`.
+-- A global symbol table keeps track of which values currently exist, ensuring we always deduplicate symbols.
+-- This therefore allows us to:
+-- - Check for equality between symbols in constant-time (using pointer equality)
+-- - Calculate the hash in constant-time (using `StableName`)
+-- - Keep the memory footprint of repeatedly-seen strings low.
+--
+-- This is very useful if you're frequently comparing strings
+-- and the same strings might come up many times.
+-- It also makes Symbol a great candidate for a key in a `HashMap` or `Data.HashSet`.
+--
+-- The global symbol table is implemented using weak pointers,
+-- which means that unused symbols will be garbage collected.
+-- As such, you do not need to be concerned about memory leaks (as is the case with many other symbol table implementations).
+--
+-- Symbols are considered 'the same' regardless of whether they originate
+-- from a `String`, (lazy or strict, normal or short) `Data.Text`, (lazy or strict, normal or short) `Data.ByteString` etc.
+data Symbol where
+    Symbol :: {-# UNPACK #-} !Symbol# -> Symbol
 
-instance Eq Symbol where
-  a == b = innerByteArray# a `Symbolize.Accursed.sameByteArray` innerByteArray# b
-  {-# INLINE (==) #-}
+-- | An unlifted version of `Symbol`;
+-- this is the actual datatype we do most calculations on.
+--
+-- Exposed to users because it can be useful to store symbols in containers with unlifted datatypes,
+-- which results in less memory usage and smaller code 
+-- (since the 'does this still need to be evaluated?' checks can be elided)
+data Symbol# :: UnliftedType where 
+    Symbol# :: ByteArray# -> Symbol#
 
-instance Ord Symbol where
-    -- We do an extra ptr-equality check here,
-    -- since ShortText doesn't do it
-    compare a b | a == b = EQ
-    compare a b = compare (symbolToShortText a) (symbolToShortText b)
-    {-# INLINE compare #-}
-
-
-instance Hashable.Hashable Symbol where
-    hash = symbolAddrInt
-    {-# INLINE hash #-}
-    hashWithSalt = Hashable.defaultHashWithSalt
-    {-# INLINE hashWithSalt #-}
-
--- instance Display Symbol where
---     displayBuilder = unintern
---     {-# INLINE displayBuilder #-}
-
+newtype WeakSymbol# = WeakSymbol# (Weak# Symbol#)
 instance Show Symbol where
-    showsPrec p symbol = 
-        let !str = unintern @String symbol
-        in showParen (p > 10) $
-            showString "Symbolize.intern " . shows str -- . showString (" " <> show (symbolAddrInt symbol))
+  showsPrec p symbol =
+    let !str = unintern @ShortText symbol
+     in showParen (p > 10) $
+          showString "Symbolize.intern " . shows str
 
--- | Symbol contains by definition always an already-evaluated ByteArray#
+-- | Equality checking takes only O(1) time, and is a simple pointer-equality check.
+instance Eq Symbol where
+    {-# INLINE (==) #-}
+    (Symbol sym1#) == (Symbol sym2#) =
+        case reallyUnsafePtrEquality# sym1# sym2# of
+            0# -> False
+            _ -> True
+-- | Symbols are ordered by their `ShortText` representation.
+--
+-- Comparison takes O(n) time, as they are compared byte-by-byte.
+instance Ord Symbol where
+    {-# INLINE compare #-}
+    a `compare` b = symbolToShortText a `compare` symbolToShortText b
+
+-- | The contents inside a `Symbol` are always guaranteed to be evaluated,
+-- so we only call `seq`.
 instance NFData Symbol where
-  rnf sym = seq sym ()
+    {-# INLINE rnf #-}
+    rnf a = seq a ()
+
+-- |
+-- Hashing a `Symbol` is very fast:
+--
+-- `hash` is a no-op and results in zero collissions, as `StableName`s are used.
+--
+-- `hashWithSalt` takes O(1) time; just as long as hashWithSalt-ing any other `Int`.
+instance Hashable Symbol where
+    {-# INLINE hash #-}
+    hash = symbolHash
+    {-# INLINE hashWithSalt #-}
+    hashWithSalt salt symbol = hashWithSalt salt (symbolHash symbol)
+
+instance IsString Symbol where
+    fromString = intern
 
 -- | To be a good citizen w.r.t both `Show` and `IsString`, reading is supported two ways:
 --
@@ -197,159 +225,147 @@ instance Read Symbol where
         str <- readPrec @String
         return $ intern str
 
-instance IsString Symbol where
-    fromString = intern
-    {-# INLINE fromString #-}
+data WeakSymbol where
+    WeakSymbol :: {-# UNPACK #-} !WeakSymbol# -> WeakSymbol
 
-innerByteArray# :: Symbol -> ByteArray#
-innerByteArray# (Symbol (ByteArray ba#)) = ba#
+newtype SymbolTable = SymbolTable {unSymbolTable :: Map ShortText WeakSymbol}
 
-symbolAddrInt :: Symbol -> Int
-symbolAddrInt s = I# (addr2Int# (byteArrayContents# (innerByteArray# s)))
-
-symbolToShortText :: Symbol -> ShortText
-symbolToShortText (Symbol (ByteArray ba#)) = 
-    (SBS ba#)
-    & Text.Short.Unsafe.fromShortByteStringUnsafe
-
-unintern :: (Textual str) => Symbol -> str
-{-# INLINE unintern #-}
-unintern !s =
-    s
-    & symbolToShortText
-    & Textual.fromShortText
-
-intern :: (Textual str) => str -> Symbol
-{-# INLINE intern #-}
-intern !str = 
-    let !text = Textual.toShortText str in
-        -- SAFETY: `intern` is idempotent, so inlining and CSE is benign (and might indeed improve performance).
-        Symbolize.Accursed.accursedUnutterablePerformIO $ do
-            -- Optimization: non-atomically see if the symbol already exists:
-            maybeSym <- lookupInner text
-            case maybeSym of
-                Just sym -> pure sym
-                Nothing -> 
-                    -- If it doesn't, time to modify the table
-                    IORef.atomicModifyIORef' (symbolTableRef globalSymbolTable') $ \symtab -> 
-                        -- SAFETY: Can never float outside of the call to `atomicModifyIORef'
-                        Symbolize.Accursed.accursedUnutterablePerformIO $ do
-                        -- We re-check if the symbol exists, as another thread might have inserted it
-                        -- in-between above check and aquiring the IORef
-                        res <- lookupCritical symtab text
-                        case res of
-                            Just sym -> pure (symtab, sym)
-                            Nothing ->
-                                insert symtab text
-
-lookup :: (Textual str) => str -> IO (Maybe Symbol)
-lookup = lookupInner . Textual.toShortText
-
-lookupInner :: ShortText -> IO (Maybe Symbol)
-lookupInner text = do
-    symtab <- IORef.readIORef (symbolTableRef globalSymbolTable')
-    lookupCritical symtab text
-
-lookupCritical :: SymbolTable -> ShortText -> IO (Maybe Symbol)
-lookupCritical (SymbolTable table) text = 
-    case setLookup (T text) table of
-        Nothing -> pure Nothing
-        Just (S (WeakSymbol weak)) -> do
-            res <- Weak.deRefWeak weak
-            pure $ (fmap Symbol) res
-        Just (T _) -> error "This should never happen; we never store T's in the symbol set"
-    where
-        setLookup elem set = 
-            set
-            & Set.lookupIndex elem
-            & fmap (\idx -> Set.elemAt idx set)
-
-
-insert :: SymbolTable -> ShortText -> IO (SymbolTable, Symbol)
-insert (SymbolTable table) text = do
-    let !ba = shortTextToByteArray text
-    let !sym = Symbolize.Accursed.ensurePinned ba
-    weak <- Symbolize.Accursed.mkWeakByteArray sym (pure ())-- (finalizer text)
-    let weak' = S (WeakSymbol weak)
-    let table' = Set.insert weak' table
-    let idx = Set.findIndex weak' table'
-    Weak.addFinalizer sym (finalizer text idx)
-    pure (SymbolTable $ table', Symbol sym)
-
-
-
-newtype WeakSymbol = WeakSymbol (Weak ByteArray)
-
--- instance Eq WeakSymbol where
---     a == b = 
---         let 
---             a' = symbolToShortText (deRefWeakSymbol a)
---             b' = symbolToShortText (deRefWeakSymbol b)
---         in
---             a' == b'
-
--- instance Ord WeakSymbol where
---     a `compare` b = deRefWeakSymbol a `compare` deRefWeakSymbol b
-
-instance Show WeakSymbol where
-    show = show . deRefWeakSymbol
-
-data WeakSymbolOrShortText = S {-# UNPACK #-} !WeakSymbol | T {-# UNPACK #-} !ShortText
-  deriving Show
-
-instance Eq WeakSymbolOrShortText where
-    a == b = normalize a == normalize b
-
-instance Ord WeakSymbolOrShortText where
-    a `compare` b = normalize a `compare` normalize b
-
-normalize (S s) = case deRefWeakSymbol s of
-    Nothing -> Nothing
-    Just s' -> Just $ symbolToShortText s'
-normalize (T t) = Just t
-
-deRefWeakSymbol :: WeakSymbol -> Maybe Symbol
-deRefWeakSymbol (WeakSymbol weak) = 
-    case (System.IO.Unsafe.unsafePerformIO (Weak.deRefWeak weak)) of
-        Nothing -> Nothing
-        Just ba -> Just (Symbol ba)
-
-
-newtype SymbolTable = SymbolTable (Set WeakSymbolOrShortText)
-
--- newtype SymbolTable = SymbolTable (Map ShortText (Weak ByteArray))
 newtype GlobalSymbolTable = GlobalSymbolTable {symbolTableRef :: IORef SymbolTable}
 
 instance Show GlobalSymbolTable where
     -- SAFETY: We're only reading, and do not care about performance here.
     show table = System.IO.Unsafe.unsafePerformIO $ do
         SymbolTable symtab <- IORef.readIORef (symbolTableRef table)
-        let contents = Set.toList symtab
+        let contents = Map.keys symtab
         pure $ "GlobalSymbolTable { contents = " <> show contents <> " }"
 
+symbolHash :: Symbol -> Int
+symbolHash (Symbol sym#) = 
+    Symbolize.Accursed.accursedUnutterablePerformIO $ IO $ \s1 ->
+        case makeStableName# sym# s1 of
+            (# s2, sname# #) -> (# s2, I# (stableNameToInt#  sname#) #)
+
+-- | Looks up a symbol in the global symbol table.
+--
+-- Returns `Nothing` if no such symbol currently exists.
+--
+-- Takes O(log2(n)) time, where n is the number of symbols currently in the table.
+--
+-- Runs concurrently with any other operation on the symbol table, without any atomic memory barriers.
+--
+-- Because the result can vary depending on the current state of the symbol table, this function is not pure.
+lookup :: Textual str => str -> IO (Maybe Symbol)
+lookup str = do
+    let !text = Textual.toShortText str
+    GlobalSymbolTable gsymtab <- globalSymbolTable
+    symtab <- IORef.readIORef gsymtab
+    pure (lookupCritical text symtab)
+
+{-# INLINE lookupCritical #-}
+lookupCritical :: ShortText -> SymbolTable -> Maybe Symbol
+lookupCritical !text (SymbolTable hashmap) = Map.lookup text hashmap >>= deRefWeakSymbol
+
+-- | Intern a string-like value.
+--
+-- First converts s to a `ShortText` (if it isn't already one). See `Textual` for the type-specific time complexity of this.
+--
+-- Then, takes O(log2(n)) time to try to look up a matching symbol 
+-- and insert it if it did not exist yet
+-- (where n is the number of symbols currently in the table).
+--
+-- Any concurrent calls to (the critical section in) `intern` are synchronized.
+intern :: Textual str => str -> Symbol
+{-# INLINABLE intern #-}
+intern !str =
+    let !text = Textual.toShortText str in
+    -- SAFETY: It is perfectly fine if floating/sharing happens
+    -- because it is 'outwardly pure'
+    -- furthermore, if the IO action is executed twice concurrently,
+    -- synchronization will hapen at the atomicModifyIORef
+    Symbolize.Accursed.accursedUnutterablePerformIO $ do
+        GlobalSymbolTable gsymtab <- globalSymbolTable
+        IORef.atomicModifyIORef' gsymtab $ \symtab ->
+            case lookupCritical text symtab of
+                Just existingSymbol -> (symtab, existingSymbol)
+                Nothing -> Symbolize.Accursed.accursedUnutterablePerformIO $ do
+                    let !newSymbol = shortTextToNewSymbol text
+                    weakSymbol <- mkWeakSymbol newSymbol (finalizer newSymbol)
+                    let symtab' = SymbolTable $ Map.insert text weakSymbol $ unSymbolTable symtab
+                    pure (symtab', newSymbol)
+
+shortTextToNewSymbol :: ShortText -> Symbol
+shortTextToNewSymbol !text = 
+    let !(SBS ba#) = Text.Short.toShortByteString text
+    in Symbol (Symbol# ba#)
+
+mkWeakSymbol :: Symbol -> IO () -> IO WeakSymbol
+mkWeakSymbol (Symbol sym#) (IO finalizer#) = do
+    IO $ \s1 ->
+        case mkWeak# sym# sym# finalizer# s1 of
+            (# s2, weak #) -> (# s2, WeakSymbol (WeakSymbol# weak) #)
+
+finalizer :: Symbol -> IO ()
+finalizer !symbol = do
+    -- putStrLn $ "Running finalizer for symbol " <> show symbol
+    let !text = symbolToShortText symbol
+    (GlobalSymbolTable gsymtab) <- globalSymbolTable
+    IORef.atomicModifyIORef' gsymtab $ \(SymbolTable symtab) -> 
+            case deRefWeakSymbol <$> Map.lookup text symtab of
+                Just Nothing ->
+                    let symtab' = Map.delete text symtab in
+                    (SymbolTable symtab', ())
+                _ -> 
+                    -- Nothing to do if (a) it already was removed,
+                    -- or (b) if it 'again exists' which means it was re-inserted
+                    -- in-between GC and this finalizer being triggered
+                    (SymbolTable symtab, ())
+
+deRefWeakSymbol :: WeakSymbol -> Maybe Symbol
+deRefWeakSymbol (WeakSymbol (WeakSymbol# w)) = Symbolize.Accursed.accursedUnutterablePerformIO $ IO $ \s ->
+   case deRefWeak# w s of
+        (# s1, flag, p #) -> case flag of
+                                0# -> (# s1, Nothing #)
+                                _  -> (# s1, Just (Symbol p) #)
+
+-- | Unintern a symbol, returning its textual value.
+-- Looking up the Symbol's textual value takes O(1) time, as we simply follow its internal pointer.
+--
+-- Afterwards, the textual value is converted to the desired type s. See `Textual` for the type-specific time complexity.
+--
+-- Does not use the symbol table, so runs fully concurrently with any other functions manipulating it.
+unintern :: Textual str => Symbol -> str
+{-# INLINE unintern #-}
+{-# SPECIALIZE unintern :: Symbol -> ShortText #-}
+unintern symbol = Textual.fromShortText (symbolToShortText symbol)
+
+symbolToShortText :: Symbol -> ShortText
+symbolToShortText (Symbol (Symbol# ba#)) =
+    Text.Short.Unsafe.fromShortByteStringUnsafe (SBS ba#)
+
+-- | The global Symbol Table, containing a mapping between each symbol's textual representation and its deduplicated pointer.
+--
+-- You cannot manipulate the table itself directly,
+-- but you can use `globalSymbolTable` to get a handle to it and use its `Show` instance for introspection.
+--
+-- `globalSymbolTableSize` can similarly be used to get the current size of the table.
+--
+-- Current implementation details (these might change even between PVP-compatible versions):
+-- 
+-- - A (containers) `Map` is used for mapping text -> symbol. This has O(log2(n)) lookup time, but is resistent to HashDoS attacks.
 globalSymbolTable :: IO GlobalSymbolTable
 globalSymbolTable = pure globalSymbolTable'
 
 globalSymbolTable' :: GlobalSymbolTable
+  -- SAFETY: We need all calls to globalSymbolTable' to use the same thunk, so NOINLINE.
+{-# NOINLINE globalSymbolTable' #-}
 globalSymbolTable' = System.IO.Unsafe.unsafePerformIO $ do
-    ref <- IORef.newIORef (SymbolTable mempty)
+    let !set = mempty
+    !ref <- IORef.newIORef (SymbolTable set)
     pure (GlobalSymbolTable ref)
 
+-- | Returns the current size of the global symbol table. Useful for introspection or metrics.
 globalSymbolTableSize :: IO Word
 globalSymbolTableSize = do
     table <- globalSymbolTable
-    (SymbolTable symtab) <- IORef.readIORef (symbolTableRef table)
-    let size = fromIntegral $ Set.size symtab
-    pure size
-
-shortTextToByteArray :: ShortText -> ByteArray
-shortTextToByteArray !text = 
-    let !(SBS ba#) = Text.Short.toShortByteString text
-    in ByteArray ba#
-
-finalizer :: ShortText -> Int -> IO ()
-finalizer key idx = do
-    -- putStrLn $ "Deleting key " <> (show key)
-    IORef.atomicModifyIORef' (symbolTableRef globalSymbolTable') $ \(SymbolTable symtab) ->
-        (SymbolTable (Set.deleteAt idx symtab), ())
-{-# NOINLINE finalizer #-}
+    (SymbolTable hashmap) <- IORef.readIORef (symbolTableRef table)
+    pure (fromIntegral (Map.size hashmap))
