+++ src/Symbolize/Accursed.hs
@@ -1,12 +1,13 @@
 {-# LANGUAGE GHC2021, MagicHash, UnboxedTuples #-}
-module Symbolize.Accursed (accursedUnutterablePerformIO, sameByteArray, ensurePinned, mkWeakByteArray) where
+module Symbolize.Accursed (accursedUnutterablePerformIO) where
 
 import GHC.IO (IO(IO))
-import Data.Primitive.ByteArray (ByteArray(..), ByteArray#)
-import Data.Primitive.ByteArray qualified as ByteArray
-import GHC.Exts (realWorld#, reallyUnsafePtrEquality, isTrue#, mkWeak#)
-import GHC.Weak (Weak(..))
-import Unsafe.Coerce (unsafeCoerce#)
+-- import Data.Primitive.ByteArray (ByteArray(..), ByteArray#)
+-- import Data.Primitive.ByteArray qualified as ByteArray
+-- import GHC.Exts (realWorld#, reallyUnsafePtrEquality, isTrue#, mkWeak#)
+import GHC.Exts (realWorld#)
+-- import GHC.Weak (Weak(..))
+-- import Unsafe.Coerce (unsafeCoerce#)
 
 -- This \"function\" has a superficial similarity to 'System.IO.Unsafe.unsafePerformIO' but
 -- it is in fact a malevolent agent of chaos.
@@ -18,28 +19,28 @@ accursedUnutterablePerformIO (IO m) = case m realWorld# of (# _, r #)   -> r
 {-# INLINE accursedUnutterablePerformIO #-}
 
 
--- | Do two byte arrays share the same pointer?
-sameByteArray :: ByteArray# -> ByteArray# -> Bool
-sameByteArray ba1 ba2 =
-    case reallyUnsafePtrEquality (unsafeCoerce# ba1 :: ()) (unsafeCoerce# ba2 :: ()) of
-      r -> isTrue# r
+-- -- | Do two byte arrays share the same pointer?
+-- sameByteArray :: ByteArray# -> ByteArray# -> Bool
+-- sameByteArray ba1 ba2 =
+--     case reallyUnsafePtrEquality (unsafeCoerce# ba1 :: ()) (unsafeCoerce# ba2 :: ()) of
+--       r -> isTrue# r
 
--- | If already pinned, returns the input unchanged.
--- Otherwise, creates a pinned copy
-ensurePinned :: ByteArray -> ByteArray
-ensurePinned ba 
-  | ByteArray.isByteArrayPinned ba = ba
-  | otherwise = accursedUnutterablePerformIO $ do
-      pinned <- ByteArray.newPinnedByteArray (ByteArray.sizeofByteArray ba)
-      ByteArray.copyByteArray pinned 0 ba 0 (ByteArray.sizeofByteArray ba)
-      ByteArray.unsafeFreezeByteArray pinned
+-- -- | If already pinned, returns the input unchanged.
+-- -- Otherwise, creates a pinned copy
+-- ensurePinned :: ByteArray -> ByteArray
+-- ensurePinned ba 
+--   | ByteArray.isByteArrayPinned ba = ba
+--   | otherwise = accursedUnutterablePerformIO $ do
+--       pinned <- ByteArray.newPinnedByteArray (ByteArray.sizeofByteArray ba)
+--       ByteArray.copyByteArray pinned 0 ba 0 (ByteArray.sizeofByteArray ba)
+--       ByteArray.unsafeFreezeByteArray pinned
 
 
--- | Make a 'Weak' pointer to an 'ByteArray'
---
--- Based on the various `mkWeak*` functions existing in `base`
--- that add a Weak pointer to a boxed unlifted type
--- such as e.g. `mkWeakMVar`.
-mkWeakByteArray :: ByteArray -> IO () -> IO (Weak ByteArray)
-mkWeakByteArray m@(ByteArray ba#) (IO f) = IO $ \s ->
-    case mkWeak# ba# m f s of (# s1, w #) -> (# s1, Weak w #)
+-- -- | Make a 'Weak' pointer to an 'ByteArray'
+-- --
+-- -- Based on the various `mkWeak*` functions existing in `base`
+-- -- that add a Weak pointer to a boxed unlifted type
+-- -- such as e.g. `mkWeakMVar`.
+-- mkWeakByteArray :: ByteArray -> IO () -> IO (Weak ByteArray)
+-- mkWeakByteArray m@(ByteArray ba#) (IO f) = IO $ \s ->
+--     case mkWeak# ba# m f s of (# s1, w #) -> (# s1, Weak w #)
