+++ symbolize.cabal
@@ -43,9 +43,6 @@ source-repository head
 library
   exposed-modules:
       Symbolize
-      Symbolize3
-      Symbolize4
-      Symbolize5
   other-modules:
       Symbolize.Textual
       Symbolize.Accursed
@@ -60,14 +57,12 @@ library
   ghc-options: -Wall -Wcompat -Widentities -Wincomplete-record-updates -Wincomplete-uni-patterns -Wmissing-export-lists -Wmissing-home-modules -Wpartial-fields -Wredundant-constraints
   build-depends:
       base >=4.7 && <5
-    , bytestring >=0.12.0 && <0.13
+    , bytestring >=0.11.0 && <0.13
     , containers >=0.6.0 && <0.7
-    , deepseq >=1.5.0 && <1.6
+    , deepseq >=1.4.0 && <1.6
     , hashable >=1.4.0 && <1.5
-    , primitive >=0.8.0 && <0.10
     , text >=2.0 && <2.2
     , text-short >=0.1.0 && <0.2
-    , unordered-containers >=0.2.0 && <0.3
   default-language: Haskell2010
 
 test-suite symbolize-doctest
@@ -88,16 +83,14 @@ test-suite symbolize-doctest
   ghc-options: -Wall -Wcompat -Widentities -Wincomplete-record-updates -Wincomplete-uni-patterns -Wmissing-export-lists -Wmissing-home-modules -Wpartial-fields -Wredundant-constraints -threaded -rtsopts -with-rtsopts=-N
   build-depends:
       base >=4.7 && <5
-    , bytestring >=0.12.0 && <0.13
+    , bytestring >=0.11.0 && <0.13
     , containers >=0.6.0 && <0.7
-    , deepseq >=1.5.0 && <1.6
+    , deepseq >=1.4.0 && <1.6
     , doctest-parallel
     , hashable >=1.4.0 && <1.5
-    , primitive >=0.8.0 && <0.10
     , symbolize
     , text >=2.0 && <2.2
     , text-short >=0.1.0 && <0.2
-    , unordered-containers >=0.2.0 && <0.3
   default-language: Haskell2010
 
 test-suite symbolize-test
@@ -122,12 +115,11 @@ test-suite symbolize-test
   build-depends:
       async
     , base >=4.7 && <5
-    , bytestring >=0.12.0 && <0.13
+    , bytestring >=0.11.0 && <0.13
     , containers >=0.6.0 && <0.7
-    , deepseq >=1.5.0 && <1.6
+    , deepseq >=1.4.0 && <1.6
     , hashable >=1.4.0 && <1.5
     , hedgehog
-    , primitive >=0.8.0 && <0.10
     , symbolize
     , tasty
     , tasty-golden
@@ -135,5 +127,4 @@ test-suite symbolize-test
     , tasty-hunit
     , text >=2.0 && <2.2
     , text-short >=0.1.0 && <0.2
-    , unordered-containers >=0.2.0 && <0.3
   default-language: Haskell2010
